requirements.txt
langchain
langchain-groq
langchain-community
langgraph
pypdf
python-dotenv
sentence-transformers